moving ssl configs
