+++ hi-ml-cpath/src/SSL/configs/HistoSimCLRContainer.py
