+++ hi-ml-cpath/src/SSL/configs/PANDA_SimCLRContainer.py
@@ -9,8 +9,8 @@ from SSL.lightning_containers.ssl_container import EncoderName, SSLContainer, SS
 from SSL.utils import SSLTrainingType
 from health_azure.utils import is_running_in_azure_ml
 from health_cpath.datasets.panda_tiles_dataset import PandaTilesDatasetWithReturnIndex
-from health_cpath.configs.SSL.HistoSimCLRContainer import HistoSSLContainer
 from health_cpath.datasets.default_paths import PANDA_TILES_DATASET_ID
+from SSL.configs.HistoSimCLRContainer import HistoSSLContainer
 
 
 class SSLDatasetNameHiml(SSLDatasetName, Enum):  # type: ignore
